pyproject.toml
[project]
name = "Chapito"
version = "0.1.12"
description = "Free API access using your web-based chatbots"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "beautifulsoup4>=4.13.3",
    "fastapi>=0.115.11",
    "pyperclip>=1.9.0",
    "requests>=2.32.3",
    "pydoll-python>=0.1.0",
    "uvicorn>=0.34.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.5",
    "pytest-asyncio",
]
